davetjet/requirements.txt
amqp==5.3.1
ansicon==1.89.0
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
blessed==1.21.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crypto==1.4.1
cryptography==45.0.6
cssbeautifier==1.15.4
Django==5.2.2
django-cors-headers==4.7.0
django-picklefield==3.3
djangorestframework==3.16.0
djangorestframework-api-key==3.1.0
djangorestframework_simplejwt==5.5.1
djlint==1.36.4
EditorConfig==0.17.1
et_xmlfile==2.0.0
idna==3.10
jinxed==1.3.0
jsbeautifier==1.15.4
json5==0.12.0
kombu==5.5.4
Naked==0.1.32
NetGsm==0.0.7
netgsm-sms==1.1.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==3.5.3
regex==2025.7.34
requests==2.32.4
resend==2.11.0
setuptools==80.9.0
shellescape==3.8.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
